Removed the old implementation, a small reduction in features but overall a cleaner api
