+++ src/main.rs
@@ -56,45 +56,15 @@ async fn main() -> io::Result<()> {
             .service(web::resource("/integrity").route(web::get().to(api::check_ledger_integrity)))
             .service(
                 web::scope("/transactions")
-                    .service(web::resource("").route(web::get().to(transaction::list_transactions)))
                     .service(
-                        web::resource("/detail")
-                            .route(web::get().to(transaction::list_transactions_with_details)),
-                    )
-                    .service(
-                        web::resource("/{year}/{month}")
-                            .route(web::get().to(transaction::list_transactions_date_scoped)),
-                    )
-                    .service(web::resource("/{year}/{month}/detail").route(
-                        web::get().to(transaction::list_transactions_date_scoped_with_details),
-                    )),
-            )
-            .service(
-                web::scope("/transaction")
-                    .service(
-                        web::resource("").route(web::post().to(transaction::create_transaction)),
+                        web::resource("")
+                            .route(web::get().to(transaction::list_transactions))
+                            .route(web::post().to(transaction::create_transaction)),
                     )
                     .service(
                         web::resource("/{id}")
                             .route(web::get().to(transaction::get_transaction))
-                            .route(web::delete().to(transaction::delete_transaction)),
-                    )
-                    .service(
-                        web::resource("/{id}/detail")
-                            .route(web::get().to(transaction::get_transaction_detail)),
-                    ),
-            )
-            .service(
-                web::scope("/v2/transactions")
-                    .service(
-                        web::resource("")
-                            .route(web::get().to(transaction::list_transactions))
-                            .route(web::post().to(transaction::create_transaction_v2)),
-                    )
-                    .service(
-                        web::resource("/{id}")
-                            .route(web::get().to(transaction::get_transaction_v2))
-                            .route(web::put().to(transaction::update_transaction_v2))
+                            .route(web::put().to(transaction::update_transaction))
                             .route(web::delete().to(transaction::delete_transaction)),
                     ),
             )
